+++ src/microsoft-dotnet-interactive/src/connection.ts
@@ -117,118 +117,6 @@ export function isArrayOfString(collection: any): collection is string[] {
     return Array.isArray(collection) && collection.length > 0 && typeof (collection[0]) === typeof ("");
 }
 
-export function createKernelUri(kernelUri: string): string {
-    kernelUri;//?
-    const uri = URI.parse(kernelUri);
-    uri.authority;//?
-    uri.path;//?
-    let absoluteUri = `${uri.scheme}://${uri.authority}${uri.path || "/"}`;
-    return absoluteUri;//?
-}
-
-export function stampCommandRoutingSlip(kernelCommandEnvelope: contracts.KernelCommandEnvelope, kernelUri: string) {
-    stampRoutingSlip(kernelCommandEnvelope, kernelUri);
-}
-
-export function stampEventRoutingSlip(kernelEventEnvelope: contracts.KernelEventEnvelope, kernelUri: string) {
-    stampRoutingSlip(kernelEventEnvelope, kernelUri);
-}
-
-function stampRoutingSlip(kernelCommandOrEventEnvelope: KernelCommandOrEventEnvelope, kernelUri: string) {
-    if (kernelCommandOrEventEnvelope.routingSlip === undefined || kernelCommandOrEventEnvelope.routingSlip === null) {
-        kernelCommandOrEventEnvelope.routingSlip = [];
-    }
-    const normalizedUri = createKernelUri(kernelUri);
-    const canAdd = !kernelCommandOrEventEnvelope.routingSlip.find(e => createKernelUri(e) === normalizedUri);
-    if (canAdd) {
-        kernelCommandOrEventEnvelope.routingSlip.push(normalizedUri);
-        kernelCommandOrEventEnvelope.routingSlip;//?
-    } else {
-        throw new Error(`The uri ${normalizedUri} is already in the routing slip [${kernelCommandOrEventEnvelope.routingSlip}]`);
-    }
-}
-
-function continueRoutingSlip(kernelCommandOrEventEnvelope: KernelCommandOrEventEnvelope, kernelUris: string[]): void {
-    if (kernelCommandOrEventEnvelope.routingSlip === undefined || kernelCommandOrEventEnvelope.routingSlip === null) {
-        kernelCommandOrEventEnvelope.routingSlip = [];
-    }
-
-    let toContinue = createRoutingSlip(kernelUris);
-
-    if (routingSlipStartsWith(toContinue, kernelCommandOrEventEnvelope.routingSlip)) {
-        toContinue = toContinue.slice(kernelCommandOrEventEnvelope.routingSlip.length);
-    }
-
-    const original = [...kernelCommandOrEventEnvelope.routingSlip];
-    for (let i = 0; i < toContinue.length; i++) {
-        const normalizedUri = toContinue[i];//?
-        const canAdd = !kernelCommandOrEventEnvelope.routingSlip.find(e => createKernelUri(e) === normalizedUri);
-        if (canAdd) {
-            kernelCommandOrEventEnvelope.routingSlip.push(normalizedUri);
-        } else {
-            throw new Error(`The uri ${normalizedUri} is already in the routing slip [${original}], cannot continue with routing slip [${kernelUris.map(e => createKernelUri(e))}]`);
-        }
-    }
-}
-
-
-export function continueCommandRoutingSlip(kernelCommandEnvelope: contracts.KernelCommandEnvelope, kernelUris: string[]): void {
-    continueRoutingSlip(kernelCommandEnvelope, kernelUris);
-}
-
-export function continueEventRoutingSlip(kernelEventEnvelope: contracts.KernelEventEnvelope, kernelUris: string[]): void {
-    continueRoutingSlip(kernelEventEnvelope, kernelUris);
-}
-
-export function createRoutingSlip(kernelUris: string[]): string[] {
-    return Array.from(new Set(kernelUris.map(e => createKernelUri(e))));
-}
-
-export function eventRoutingSlipStartsWith(thisEvent: contracts.KernelEventEnvelope, other: string[] | contracts.KernelEventEnvelope): boolean {
-    const thisKernelUris = thisEvent.routingSlip ?? [];
-    const otherKernelUris = (other instanceof Array ? other : other?.routingSlip) ?? [];
-
-    return routingSlipStartsWith(thisKernelUris, otherKernelUris);
-}
-
-export function commandRoutingSlipStartsWith(thisCommand: contracts.KernelCommandEnvelope, other: string[] | contracts.KernelCommandEnvelope): boolean {
-    const thisKernelUris = thisCommand.routingSlip ?? [];
-    const otherKernelUris = (other instanceof Array ? other : other?.routingSlip) ?? [];
-
-    return routingSlipStartsWith(thisKernelUris, otherKernelUris);
-}
-
-
-function routingSlipStartsWith(thisKernelUris: string[], otherKernelUris: string[]): boolean {
-    let startsWith = true;
-
-    if (otherKernelUris.length > 0 && thisKernelUris.length >= otherKernelUris.length) {
-        for (let i = 0; i < otherKernelUris.length; i++) {
-            if (createKernelUri(otherKernelUris[i]) !== createKernelUri(thisKernelUris[i])) {
-                startsWith = false;
-                break;
-            }
-        }
-    }
-    else {
-        startsWith = false;
-    }
-
-    return startsWith;
-}
-
-export function eventRoutingSlipContains(kernlEvent: contracts.KernelEventEnvelope, kernelUri: string): boolean {
-    return routingSlipContains(kernlEvent, kernelUri);
-}
-
-export function commandRoutingSlipContains(kernlEvent: contracts.KernelCommandEnvelope, kernelUri: string): boolean {
-    return routingSlipContains(kernlEvent, kernelUri);
-}
-
-function routingSlipContains(kernelCommandOrEventEnvelope: KernelCommandOrEventEnvelope, kernelUri: string) {
-    const normalizedUri = createKernelUri(kernelUri);
-    return kernelCommandOrEventEnvelope?.routingSlip?.find(e => normalizedUri === createKernelUri(e)) !== undefined;
-}
 
 export function ensureOrUpdateProxyForKernelInfo(kernelInfoProduced: contracts.KernelInfoProduced, compositeKernel: CompositeKernel) {
     const uriToLookup = kernelInfoProduced.kernelInfo.remoteUri ?? kernelInfoProduced.kernelInfo.uri;
