+++ src/microsoft-dotnet-interactive/src/index.ts
@@ -15,5 +15,6 @@ export * from './kernelScheduler';
 export * from './logger';
 export * from './promiseCompletionSource';
 export * from './proxyKernel';
+export * from './routingslip';
 export * from './setup';
 export * from './tokenGenerator';
