+++ src/microsoft-dotnet-interactive/tests/routingslip.test.ts
@@ -3,7 +3,7 @@
 
 import { expect } from "chai";
 import * as contracts from "../src/contracts";
-import * as connection from "../src/connection";
+import * as routingSlip from "../src/routingslip";
 
 describe("kernel event routingSlip", () => {
     it("cannot stamp twice", () => {
@@ -13,9 +13,9 @@ describe("kernel event routingSlip", () => {
             event: {}
         };
 
-        connection.stampEventRoutingSlip(envelope, "kernel://a");
+        routingSlip.stampEventRoutingSlip(envelope, "kernel://a");
 
-        expect(() => connection.stampEventRoutingSlip(envelope, "kernel://a")).to.throw("The uri kernel://a/ is already in the routing slip [kernel://a/]");
+        expect(() => routingSlip.stampEventRoutingSlip(envelope, "kernel://a")).to.throw("The uri kernel://a/ is already in the routing slip [kernel://a/]");
     });
 
     it("can append a routing slip to another", () => {
@@ -23,12 +23,12 @@ describe("kernel event routingSlip", () => {
         let envelope: contracts.KernelEventEnvelope = {
             eventType: contracts.CommandSucceededType,
             event: {},
-            routingSlip: connection.createRoutingSlip(["kernel://a", "kernel://b"])
+            routingSlip: routingSlip.createRoutingSlip(["kernel://a", "kernel://b"])
         };
 
         let other = ["kernel://c", "kernel://d"];
 
-        connection.continueEventRoutingSlip(envelope, other);
+        routingSlip.continueEventRoutingSlip(envelope, other);
 
         expect(envelope.routingSlip).to.deep.equal(['kernel://a/', 'kernel://b/', 'kernel://c/', 'kernel://d/']);
     });
@@ -38,12 +38,12 @@ describe("kernel event routingSlip", () => {
         let envelope: contracts.KernelEventEnvelope = {
             eventType: contracts.CommandSucceededType,
             event: {},
-            routingSlip: connection.createRoutingSlip(["kernel://a", "kernel://b"])
+            routingSlip: routingSlip.createRoutingSlip(["kernel://a", "kernel://b"])
         };
 
         let other = ["kernel://a", "kernel://b", "kernel://c", "kernel://d"];
 
-        connection.continueEventRoutingSlip(envelope, other);
+        routingSlip.continueEventRoutingSlip(envelope, other);
 
         expect(envelope.routingSlip).to.deep.equal(['kernel://a/', 'kernel://b/', 'kernel://c/', 'kernel://d/']);
     });
@@ -53,10 +53,10 @@ describe("kernel event routingSlip", () => {
         let envelope: contracts.KernelEventEnvelope = {
             eventType: contracts.CommandSucceededType,
             event: {},
-            routingSlip: connection.createRoutingSlip(["kernel://a", "kernel://b"])
+            routingSlip: routingSlip.createRoutingSlip(["kernel://a", "kernel://b"])
         };
 
         let other = ["kernel://a", "kernel://c", "kernel://d", "kernel://e"];
-        expect(() => connection.continueEventRoutingSlip(envelope, other)).to.throw("The uri kernel://a/ is already in the routing slip [kernel://a/,kernel://b/], cannot continue with routing slip [kernel://a/,kernel://c/,kernel://d/,kernel://e/]");
+        expect(() => routingSlip.continueEventRoutingSlip(envelope, other)).to.throw("The uri kernel://a/ is already in the routing slip [kernel://a/,kernel://b/], cannot continue with routing slip [kernel://a/,kernel://c/,kernel://d/,kernel://e/]");
     });
 });
